refactor: move {Stateful,}Widget{,Ref} types into individual files
